4.21

+++ part_04/controllers/blog.js
@@ -38,8 +38,24 @@ blogRouter.post("/", async (request, response, next) => {
 });
 
 blogRouter.delete("/:id", async (request, response, next) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+
+  if (!user) {
+    return response.status(400).json({ error: "userId missing or not valid" });
+  }
+
+  const blog = await Blog.findById(request.params.id);
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(403).json({ error: "not authorized" });
+  }
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  }
 });
 
 blogRouter.put("/:id", async (request, response) => {
